Designed a test cse verifying that the design of #4 as proposed in be2e684 was broken
Due to the fact that analysis information is retained internally in psalm, multiple mutations
on the same file lead to conflicting file symbol declarations (repeated classes, functions) when
analyzed in sequence by the same analyzer.

This test highlights the failure by showing that an analysis on the same mutant, when performed
twice (with different mutant paths, but same code), will lead to a killed mutant due to invalid
static analysis reports.

Ref: https://github.com/Roave/infection-static-analysis-plugin/pull/196#issuecomment-840800853
Ref: https://github.com/Roave/infection-static-analysis-plugin/pull/196#issuecomment-840801843
Ref: https://github.com/Roave/infection-static-analysis-plugin/pull/196#issuecomment-840868549