Designed a failing test demonstrating that mutations collide with their source files (in definition)
Mutated files are generally a copy-cat of the original source they came from: if that source was already
somehow included by `vimeo/psalm`, the mutation would be in direct conflict with it, if it declared static
symbols such as classes or functions.

This test highlights the issue, showing that we somehow still leak shared state across mutation analysis
when multiple source files and multiple mutations on separate files are at work.