Re-implemented bootstrapping to comply with `infection/infection:0.31.9`
I'm not proud of these changes: we had to shoehorn `RunStaticAnalysisAgainstEscapedMutant`
as a child class of `TestFrameworkMutantExecutionResultFactory`, because Infection now uses
a separate `MutantExecutionResultFactory` implementation (`PHPStanMutantExecutionResultFactory`)
to run PHPStan tests.

The `PHPStanMutantExecutionResultFactory` is built in following trace (backwards):

* `PHPStanMutantExecutionResultFactory`
* `PHPStanAdapterFactory`
* `StaticAnalysisToolFactory`
* `Container#getStaticAnalysisToolFactory()`

Even if I found a way to create a clean `StaticAnalysisToolFactory`, I'd still need to override the DIC, so
for now, we stick to running this tool as it did before, hijacking the `TestFrameworkMutantExecutionResultFactory`
directly.

It is indeed going to be cleaner if we can produce a dedicated `MutantExecutionResultFactory`, which
produces its own SA-friendly `MutantExecutionResult`, instead of operating via decoration.

This was also briefly discussed on Discord: https://discord.com/channels/1090672946370052166/1091081297759309926/1442902731588374599